refactor(website): Reorganizes form components
