+++ website/src/components/forms/library/FormDetails.astro
@@ -1,8 +1,8 @@
 ---
-import type { FormSchema } from "@/components/FormMappingPlayground/types";
+import type { FormSchema } from "@/lib/types";
 import { Code } from "@astrojs/starlight/components";
 import { Tabs, TabItem } from "@astrojs/starlight/components";
-import { JsonFormRenderer } from "@/components/FormMappingPlayground/JsonFormRenderer";
+import { JsonFormRenderer } from "@/components/forms/JsonFormRenderer";
 import type { JsonSchema } from "@jsonforms/core";
 
 interface Props {
