+++ website/src/components/forms/library/FormSummary.astro
@@ -1,5 +1,5 @@
 ---
-import type { FormSchema } from "@/components/FormMappingPlayground/types";
+import type { FormSchema } from "@/lib/types";
 import type { JsonSchema } from "@jsonforms/core";
 import formsIndex from "@/content/forms/index.json";
