+++ website/src/pages/forms/[slug].astro
@@ -1,5 +1,5 @@
 ---
-import { schemas } from "@/components/FormMappingPlayground/schemas";
+import { schemas } from "@/lib/schemas";
 import StarlightPage from "@astrojs/starlight/components/StarlightPage.astro";
 import type { GetStaticPaths } from "astro";
 import FormDetails from "@/components/forms/library/FormDetails.astro";
