+++ website/src/pages/forms/library.astro
@@ -1,5 +1,5 @@
 ---
-import { schemas } from "@/components/FormMappingPlayground/schemas";
+import { schemas } from "@/lib/schemas";
 import StarlightPage from "@astrojs/starlight/components/StarlightPage.astro";
 import { CardGrid, LinkCard } from "@astrojs/starlight/components";
