+++ website/src/pages/forms/playground.astro
@@ -1,5 +1,5 @@
 ---
-import FormMappingPlayground from "@/components/FormMappingPlayground";
+import FormMappingPlayground from "@/components/forms/playground";
 import StarlightPage from "@astrojs/starlight/components/StarlightPage.astro";
 ---
